requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
aiofiles==23.1.0
google-genai==1.46.0
redis>=5.0.0
python-dotenv==1.0.0
